+++ example/lib/main.dart
@@ -14,6 +14,19 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
       ),
       home: MyHomePage(),
+      onGenerateRoute: (settings) {
+        switch (settings.name) {
+          case '/second':
+            return PageTransition(
+              child: SecondPage(),
+              type: PageTransitionType.scale,
+              settings: settings,
+            );
+            break;
+          default:
+            return null;
+        }
+      },
     );
   }
 }
@@ -35,19 +48,24 @@ class MyHomePage extends StatelessWidget {
               child: Text('Fade Second Page'),
               onPressed: () {
                 Navigator.push(
-                    context,
-                    PageTransition(
-                        type: PageTransitionType.fade, child: SecondPage()));
+                  context,
+                  PageTransition(
+                    type: PageTransitionType.fade,
+                    child: SecondPage(),
+                  ),
+                );
               },
             ),
             RaisedButton(
               child: Text('Left To Right Slide Second Page'),
               onPressed: () {
                 Navigator.push(
-                    context,
-                    PageTransition(
-                        type: PageTransitionType.leftToRight,
-                        child: SecondPage()));
+                  context,
+                  PageTransition(
+                    type: PageTransitionType.leftToRight,
+                    child: SecondPage(),
+                  ),
+                );
               },
             ),
             RaisedButton(
@@ -100,12 +118,8 @@ class MyHomePage extends StatelessWidget {
             RaisedButton(
               child: Text('Down to Up Second Page'),
               onPressed: () {
-                Navigator.push(
-                    context,
-                    PageTransition(
-                        curve: Curves.linear,
-                        type: PageTransitionType.downToUp,
-                        child: SecondPage()));
+                Navigator.pushNamed(context, "/second",
+                    arguments: "with Arguments");
               },
             ),
           ],
@@ -116,11 +130,15 @@ class MyHomePage extends StatelessWidget {
 }
 
 class SecondPage extends StatelessWidget {
+  final String title;
+
+  const SecondPage({Key key, this.title}) : super(key: key);
   @override
   Widget build(BuildContext context) {
+    final args = ModalRoute.of(context).settings.arguments;
     return Scaffold(
       appBar: AppBar(
-        title: Text('Page Transition SecondPage'),
+        title: Text(args ?? "Page Transition Plugin"),
       ),
       body: Center(
         child: Text('Second Page'),